requirements.txt
PyQt5>=5.15.9
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.12.1
requests>=2.31.0
pandas>=2.1.0
pytest>=7.4.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pillow>=10.0.0
openpyxl>=3.1.2
pymysql>=1.1.0
dbutils>=3.0.3 